// models/saddle.model.ts
export type SaddleState = {
  data: {
    sadd_type: string;
    sadd_finished_size: string;
    sadd_cover: string;
    sadd_text: string;
    sadd_cover_paper: string;
    sadd_text_paper: string;
    sadd_printing: string;
    sadd_cover_coating: string;
    sadd_text_coating: string;
    sadd_500: number;
    sadd_1000: number;
    sadd_2000: number;
    sadd_3000: number;
    sadd_4000: number;
    sadd_5000: number;
  };
  data_text_no: {
    sadd_type: string;
    sadd_finished_size: string;
    sadd_cover: string;
    sadd_text: string;
    sadd_cover_paper: string;
    sadd_text_paper: string;
    sadd_printing: string;
    sadd_cover_coating: string;
    sadd_text_coating: string;
    sadd_500: number;
    sadd_1000: number;
    sadd_2000: number;
    sadd_3000: number;
    sadd_4000: number;
    sadd_5000: number;
  };
  data_text_paper: {
    sadd_type: string;
    sadd_finished_size: string;
    sadd_cover: string;
    sadd_text: string;
    sadd_cover_paper: string;
    sadd_text_paper: string;
    sadd_printing: string;
    sadd_cover_coating: string;
    sadd_text_coating: string;
    sadd_500: number;
    sadd_1000: number;
    sadd_2000: number;
    sadd_3000: number;
    sadd_4000: number;
    sadd_5000: number;
  };
  data_printing: {
    sadd_type: string;
    sadd_finished_size: string;
    sadd_cover: string;
    sadd_text: string;
    sadd_cover_paper: string;
    sadd_text_paper: string;
    sadd_printing: string;
    sadd_cover_coating: string;
    sadd_text_coating: string;
    sadd_500: number;
    sadd_1000: number;
    sadd_2000: number;
    sadd_3000: number;
    sadd_4000: number;
    sadd_5000: number;
  };
  data_cover_paper_edit : {
    sadd_type: string;
    sadd_finished_size: string;
    sadd_cover: string;
    sadd_text: string;
    sadd_cover_paper: string;
    sadd_text_paper: string;
    sadd_printing: string;
    sadd_cover_coating: string;
    sadd_text_coating: string;
    sadd_500: number;
    sadd_1000: number;
    sadd_2000: number;
    sadd_3000: number;
    sadd_4000: number;
    sadd_5000: number; 
  }
};